Split MiniCanvas Component
